requirements.txt
accelerate
diffusers
transformers
numpy
pillow
matplotlib
wandb
safetensors
einops
pytorch-fid
tqdm
opencv-python-headless
peft